requirements.txt
# Core dependencies - compatible versions for Python 3.11
flask>=2.3.0,<3.1.0
flask-cors>=4.0.0,<4.1.0
streamlit>=1.27.0,<1.30.0
requests>=2.30.0,<2.32.0
python-dotenv>=0.19.0,<1.1.0
tqdm>=4.66.0,<4.67.0
gradio-client>=0.10.0,<1.0.0

# Scientific computing - COMPATIBLE VERSIONS
numpy==1.25.2
pandas==2.0.3
scipy==1.11.4
scikit-learn==1.3.2

# NLP and text processing
nltk>=3.8.0,<3.10.0

# Arabic text processing
arabic-reshaper>=3.0.0,<3.1.0
python-bidi>=0.4.2,<0.5.0

# Language detection
langdetect>=1.0.9,<1.2.0

# Visualization and explainability
plotly>=5.15.0,<5.18.0
matplotlib>=3.7.0,<3.8.0
seaborn>=0.12.0,<0.13.0

# Testing
pytest>=7.4.0,<7.5.0
pytest-cov>=4.1.0,<4.2.0

# Optional ML dependencies (install separately if needed)
# transformers>=4.30.0,<4.40.0
# torch>=2.0.0,<2.2.0
# lime>=0.2.0.1
# wordcloud>=1.9.0,<1.10.0
# textblob>=0.17.0,<0.18.0